external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/attribs/gl-vertex-attrib-expected.txt

This test ensures WebGL implementations vertexAttrib can be set and read.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Canvas.getContext
PASS context exists

Checking gl.vertexAttrib.
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[3] is 4
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[0] is 5
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[0] is 7
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[1] is 8
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[2] is 9
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[2] is 8
PASS gl.getVertexAttrib(0, gl.CURRENT_VERTEX_ATTRIB)[3] is 9
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[3] is 4
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[0] is 5
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[0] is 7
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[1] is 8
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[2] is 9
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[2] is 8
PASS gl.getVertexAttrib(1, gl.CURRENT_VERTEX_ATTRIB)[3] is 9
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[3] is 4
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[0] is 5
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[0] is 7
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[1] is 8
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[2] is 9
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[2] is 8
PASS gl.getVertexAttrib(2, gl.CURRENT_VERTEX_ATTRIB)[3] is 9
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[3] is 4
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[0] is 5
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[0] is 7
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[1] is 8
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[2] is 9
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[2] is 8
PASS gl.getVertexAttrib(3, gl.CURRENT_VERTEX_ATTRIB)[3] is 9
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[3] is 4
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[0] is 5
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[0] is 7
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[1] is 8
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[2] is 9
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[2] is 8
PASS gl.getVertexAttrib(4, gl.CURRENT_VERTEX_ATTRIB)[3] is 9
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[3] is 4
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[0] is 5
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[0] is 7
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[1] is 8
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[2] is 9
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[2] is 8
PASS gl.getVertexAttrib(5, gl.CURRENT_VERTEX_ATTRIB)[3] is 9
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[3] is 4
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[0] is 5
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[0] is 7
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[1] is 8
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[2] is 9
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[2] is 8
PASS gl.getVertexAttrib(6, gl.CURRENT_VERTEX_ATTRIB)[3] is 9
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[3] is 4
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[0] is 5
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[0] is 7
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[1] is 8
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[2] is 9
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[2] is 8
PASS gl.getVertexAttrib(7, gl.CURRENT_VERTEX_ATTRIB)[3] is 9
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[3] is 4
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[0] is 5
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[0] is 7
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[1] is 8
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[2] is 9
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[2] is 8
PASS gl.getVertexAttrib(8, gl.CURRENT_VERTEX_ATTRIB)[3] is 9
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[3] is 4
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[0] is 5
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[0] is 7
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[1] is 8
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[2] is 9
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[2] is 8
PASS gl.getVertexAttrib(9, gl.CURRENT_VERTEX_ATTRIB)[3] is 9
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[3] is 4
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[0] is 5
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[0] is 7
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[1] is 8
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[2] is 9
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[2] is 8
PASS gl.getVertexAttrib(10, gl.CURRENT_VERTEX_ATTRIB)[3] is 9
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[3] is 4
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[0] is 5
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[0] is 7
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[1] is 8
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[2] is 9
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[2] is 8
PASS gl.getVertexAttrib(11, gl.CURRENT_VERTEX_ATTRIB)[3] is 9
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[3] is 4
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[0] is 5
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[0] is 7
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[1] is 8
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[2] is 9
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[2] is 8
PASS gl.getVertexAttrib(12, gl.CURRENT_VERTEX_ATTRIB)[3] is 9
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[3] is 4
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[0] is 5
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[0] is 7
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[1] is 8
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[2] is 9
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[2] is 8
PASS gl.getVertexAttrib(13, gl.CURRENT_VERTEX_ATTRIB)[3] is 9
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[3] is 4
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[0] is 5
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[0] is 7
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[1] is 8
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[2] is 9
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[2] is 8
PASS gl.getVertexAttrib(14, gl.CURRENT_VERTEX_ATTRIB)[3] is 9
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[0] is 1
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[1] is 2
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[2] is 3
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[3] is 4
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[0] is 5
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[1] is 0
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[2] is 0
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[0] is 7
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[1] is 8
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[2] is 9
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[3] is 1
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[0] is 6
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[1] is 7
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[2] is 8
PASS gl.getVertexAttrib(15, gl.CURRENT_VERTEX_ATTRIB)[3] is 9
PASS getError was expected value: NO_ERROR : 

PASS successfullyParsed is true

TEST COMPLETE
